Mark tests as ported to the TCK
- Remove tests with only result assertions
- Remove test assertions on remaining tests
- Rename test class to clarify new purpose
